Added Alarm Script
So, this script uses at to schedule the actual alarm.
It runs the script in an xterm, and allows you to set the alarm intially
to anything 'at' recognizes, but then after that it snoozes for some
number of minutes.

That goes forever.
To stop it from snoozing, use 'atq' and 'atrm' to remove the job, so it
doesn't come up again.

Because I open a new terminal to run this in, I've added a prompt to
alarm.hs so if I read it, I know what's on my screen.