Tweaks to set_alarm
Only blank screen if we turned it on.
While I thought I wanted the feature that when I set the alarm the
screen would turn itself off, I don't.

I want to turn the screen off, but not right away, so I'll do it
manually.
I likely want to switch to an empty tab, or finish some stuff up first.

Also, I set the snooze to 10 minutes.
Real time.